include: linux: pwm: Convert phase to u64
This change is a follow-up to
commit 4fa760fa62a9 ("pwm: Convert period and duty cycle to u64"),
converting the remaining parameters to u64.

+++ include/linux/pwm.h
@@ -41,7 +41,7 @@ enum pwm_polarity {
  */
 struct pwm_args {
 	u64 period;
-	unsigned int phase;
+	u64 phase;
 	enum pwm_polarity polarity;
 };
 
@@ -61,7 +61,7 @@ enum {
 struct pwm_state {
 	u64 period;
 	u64 duty_cycle;
-	unsigned int phase;
+	u64 phase;
 	enum pwm_polarity polarity;
 	bool enabled;
 };
@@ -141,13 +141,13 @@ static inline u64 pwm_get_duty_cycle(const struct pwm_device *pwm)
 	return state.duty_cycle;
 }
 
-static inline void pwm_set_phase(struct pwm_device *pwm, unsigned int phase)
+static inline void pwm_set_phase(struct pwm_device *pwm, u64 phase)
 {
 	if (pwm)
 		pwm->state.phase = phase;
 }
 
-static inline unsigned int pwm_get_phase(const struct pwm_device *pwm)
+static inline u64 pwm_get_phase(const struct pwm_device *pwm)
 {
 	struct pwm_state state;
 
@@ -329,7 +329,7 @@ struct pwm_chip {
 struct pwm_capture {
 	u64 period;
 	u64 duty_cycle;
-	unsigned int phase;
+	u64 phase;
 };
 
 #if IS_ENABLED(CONFIG_PWM)
